requirements.txt
# HTTP
requests>=2.31.0

# Data
pandas>=2.0.0

# Config
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
